Added phone and DOB support to the model/controller, modified signup page to support these.
Likely going to be cleaning this up some as the night goes on, but I'm pushing these
changes to demonstrate them after a successful signup test. I still need
to implement error-checking for the date and phone. (Phone numbers
should be raw string of numbers, pre-emptive date checking before date
variable is placed into newUser, and any other changes that may be
important considerations that I or others find)
There is an additional "admin" variable as per one of the
recommendations on Trello, when signing up as a user this is
automatically set to false.